refactor: simplify implementation
