+++ src/components/Icon.tsx
@@ -1,69 +1,14 @@
 import * as React from 'react';
 import {
   Image,
-  Text,
-  StyleSheet,
   I18nManager,
   Platform,
   ImageSourcePropType,
 } from 'react-native';
-import { Consumer as SettingsConsumer } from '../core/settings';
-
-let MaterialCommunityIcons: any;
-
-try {
-  // Optionally require vector-icons
-  MaterialCommunityIcons = require('react-native-vector-icons/MaterialCommunityIcons')
-    .default;
-} catch (e) {
-  if (
-    // @ts-ignore
-    global.__expo &&
-    // @ts-ignore
-    global.__expo.Icon &&
-    // @ts-ignore
-    global.__expo.Icon.MaterialCommunityIcons
-  ) {
-    // Snack doesn't properly bundle vector icons from subpath
-    // Use icons from the __expo global if available
-    // @ts-ignore
-    MaterialCommunityIcons = global.__expo.Icon.MaterialCommunityIcons;
-  } else {
-    let isErrorLogged = false;
-
-    // Fallback component for icons
-    // @ts-ignore
-    MaterialCommunityIcons = ({ name, color, size, ...rest }) => {
-      /* eslint-disable no-console */
-      if (!isErrorLogged) {
-        if (
-          !/(Cannot find module|Module not found|Cannot resolve module)/.test(
-            e.message
-          )
-        ) {
-          console.error(e);
-        }
-
-        console.warn(
-          `Tried to use the icon '${name}' in a component from 'react-native-paper', but 'react-native-vector-icons' could not be loaded.`,
-          `To remove this warning, try installing 'react-native-vector-icons' or use another method to specify icon: https://callstack.github.io/react-native-paper/icons.html.`
-        );
-
-        isErrorLogged = true;
-      }
-
-      return (
-        <Text
-          {...rest}
-          style={[styles.icon, { color, fontSize: size }]}
-          pointerEvents="none"
-        >
-          □
-        </Text>
-      );
-    };
-  }
-}
+import {
+  Consumer as SettingsConsumer,
+  accessibilityProps,
+} from '../core/settings';
 
 type IconSourceBase = string | ImageSourcePropType;
 
@@ -131,17 +76,6 @@ const Icon = ({ source, color, size, ...rest }: Props) => {
       ? source.source
       : source;
 
-  const accessibilityProps =
-    Platform.OS === 'web'
-      ? {
-          role: 'img',
-          focusable: false,
-        }
-      : {
-          accessibilityElementsHidden: true,
-          importantForAccessibility: 'no-hide-descendants' as 'no-hide-descendants',
-        };
-
   if (isImageSource(s)) {
     return (
       <Image
@@ -166,34 +100,12 @@ const Icon = ({ source, color, size, ...rest }: Props) => {
     return (
       <SettingsConsumer>
         {({ icon }) => {
-          return icon ? (
-            icon({
-              name: s,
-              color,
-              size,
-              style: {
-                transform: [{ scaleX: direction === 'rtl' ? -1 : 1 }],
-                ...styles.icon,
-              },
-              pointerEvents: 'none',
-              accessibilityProps,
-            })
-          ) : (
-            <MaterialCommunityIcons
-              {...rest}
-              name={s}
-              color={color}
-              size={size}
-              style={[
-                {
-                  transform: [{ scaleX: direction === 'rtl' ? -1 : 1 }],
-                },
-                styles.icon,
-              ]}
-              pointerEvents="none"
-              {...accessibilityProps}
-            />
-          );
+          return icon({
+            name: s,
+            color,
+            size,
+            direction,
+          });
         }}
       </SettingsConsumer>
     );
@@ -205,9 +117,3 @@ const Icon = ({ source, color, size, ...rest }: Props) => {
 };
 
 export default Icon;
-
-const styles = StyleSheet.create({
-  icon: {
-    backgroundColor: 'transparent',
-  },
-});
