+++ src/core/Provider.tsx
@@ -1,6 +1,10 @@
 import * as React from 'react';
 import { ThemeProvider } from './theming';
-import { Provider as SettingsProvider, Settings } from './settings';
+import {
+  Provider as SettingsProvider,
+  Settings,
+  defaultIcon,
+} from './settings';
 import PortalHost from '../components/Portal/PortalHost';
 import { Theme } from '../types';
 
@@ -14,9 +18,7 @@ export default class Provider extends React.Component<Props> {
   render() {
     return (
       <PortalHost>
-        <SettingsProvider
-          value={this.props.settings ? this.props.settings : {}}
-        >
+        <SettingsProvider value={this.props.settings || { icon: defaultIcon }}>
           <ThemeProvider theme={this.props.theme}>
             {this.props.children}
           </ThemeProvider>
